requirements.txt
openclaw>=1.0.0
langchain>=0.4.0
langchain-google-genai>=2.0.0
openai>=1.75.0
mcp>=1.0.0
fastmcp>=0.9.0
PyYAML>=7.0
python-dotenv>=1.1.0
requests>=2.33.0
pytest>=8.5.0
